Replace youtube example with wikipedia article word count example
The youtube example used an API that is now 410 Gone.
